# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Australian SOCI Act Compliance Automation Platform
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /assets:
    get:
      summary: List assets
      responses:
        '200':
          description: A list of assets.
    post:
      summary: Create asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetCreate'
      responses:
        '201':
          description: Created asset
  /assets/bulk_import:
    post:
      summary: Bulk import assets via CSV upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                csv:
                  type: string
                  description: CSV payload using AssetCreate headers
      responses:
        '200':
          description: Result summary
  /cirmp/templates:
    get:
      summary: List templates
      responses:
        '200':
          description: CIRMP templates
  /cirmp/templates/generate:
    post:
      summary: Generate CIRMP template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateRequest'
      responses:
        '200':
          description: Generated template
  /compliance/score:
    get:
      summary: Organisation compliance score
      responses:
        '200':
          description: Score object
  /compliance/score/{asset_id}:
    get:
      summary: Asset compliance score
      parameters:
        - in: path
          name: asset_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Score object
        '404':
          description: Asset not found
  /incidents:
    post:
      summary: Create incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Incident created
  /incidents/{incident_id}/report:
    post:
      summary: Generate ACSC payload for incident
      parameters:
        - in: path
          name: incident_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: ACSC payload
  /supplychain/ingest:
    post:
      summary: Ingest supplier data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierIngestRequest'
      responses:
        '200':
          description: Supplier added
  /supplychain/risk-score:
    get:
      summary: Aggregate supplier risk
      responses:
        '200':
          description: Risk score
  /reports/board:
    post:
      summary: Generate board level report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardReportRequest'
      responses:
        '200':
          description: Board report
  /demo/seed:
    post:
      summary: Seed or refresh demo data
      parameters:
        - in: query
          name: force
          schema:
            type: boolean
            default: false
          description: Clear demo data before reseeding
      responses:
        '200':
          description: Counts per collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeedStats'
components:
  schemas:
    AssetCreate:
      type: object
      required: [name, sector, category, asset_type, criticality, owner]
      properties:
        name:
          type: string
        sector:
          type: string
        category:
          type: string
        asset_type:
          type: string
        criticality:
          type: string
        owner:
          type: string
    TemplateRequest:
      type: object
      required: [asset_type, organisation_size, sector]
      properties:
        asset_type:
          type: string
        organisation_size:
          type: string
          enum: [small, medium, large]
        sector:
          type: string
    IncidentCreate:
      type: object
      required: [title, severity, description]
      properties:
        asset_id:
          type: string
          nullable: true
        title:
          type: string
        severity:
          type: string
        description:
          type: string
        evidence_bundle:
          type: array
          items:
            type: string
    SupplierIngestRequest:
      type: object
      required: [name, geo]
      properties:
        name:
          type: string
        geo:
          type: string
        tier:
          type: integer
          default: 1
        dependencies:
          type: array
          items:
            type: string
    BoardReportRequest:
      type: object
      properties:
        audience:
          type: string
        include_supply_chain:
          type: boolean
        include_incident_summary:
          type: boolean
    SeedStats:
      type: object
      properties:
        assets:
          type: integer
        templates:
          type: integer
        suppliers:
          type: integer
        incidents:
          type: integer